Foundation laid for Fortune mechanic
Finally added ability for players to have more stats added to them as
each player is reinitialized on startup.

Gave players the luck attribute which multiplies returns on yen gets
currently. Probably will also use for combat. Combat needs to be fixed
but whatever.

Because of how players are initialized, if a player is dead when Tsumomo
boots, it will give them 15HP instead of keeping them at zero. Needs
fixing.